Create README.md
TeamCodeRed- Hack36 4.0
Welcome all! Our project for this years Hack36 is based on the theme LIFE IN PANDEMIC. The basic idea behind the project was to create an e-wallet for users with a daily set limit for spending amount. This would allow them both contactless and safe payments in this pandemic situation as well as to keep a check on their daily expenditure, as we all know how hard it is to stop ourselves from overspending our money on stuff we will probably never use or ordering food online just because we are sitting at our homes being bored! 
We really enjoyed building this project! And we hope you like it too.